replace removeRecipes with removeAllRecipesExceptFor
Each time the Recipe Iterator is retrieved, a new set of Recipe objects
are generated - so unless we cache a list and use that to re-add, the
old API methods simply will not function.

Opted to instead just have a singular method that accepts recipes to
re-add after removing instead of multiple variants of removing recipes.