More javadocs, removed internal item versioning
Versioning was going to allow for multiple items with the same name to
"exist," in that an older version of an item would still be able to be
recognized if need be via plugins. However, plugins can implement this
themselves in the string id, so we can save a call to String#split every
time we check for and compare items.

- Also removed delayed registration of command; I meant to include the
instantiation of the class in the runnable as well, but then decided
that it would be of little use; and if players would like to override
recipes set by other plugins, they are free to do so.